website_discovery/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "website-discovery"
version = "1.0.0"
description = "Persistent URL Discovery Service for Bangladeshi Government Websites"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "URL Discovery Engine Team"}
]
keywords = ["crawler", "discovery", "gov.bd", "postgresql", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing Creation",
]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.19.0",
    "uvicorn>=0.25.0",
    "fastapi>=0.110.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "nox>=2023.04.22",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
website-discovery = "src.main:main"
ingest-seeds = "src.tools.ingest_seed_urls:main"
status-report = "src.tools.status_report:main"

[project.urls]
Homepage = "https://github.com/your-org/website-discovery"
Documentation = "https://github.com/your-org/website-discovery/tree/main/docs"
Repository = "https://github.com/your-org/website-discovery"
Issues = "https://github.com/your-org/website-discovery/issues"

# ============================================
# Black Configuration
# ============================================
[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

# ============================================
# Ruff Configuration
# ============================================
[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "B904", # raise without from inside except (sometimes intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

# ============================================
# MyPy Configuration
# ============================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ============================================
# Pytest Configuration
# ============================================
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v -x --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================
# Coverage Configuration
# ============================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"
